+++ packages/patternfly-4/react-core/src/components/Dropdown/index.js
@@ -1,5 +1,5 @@
 export { default as Dropdown } from './Dropdown';
-export { DropdownDirection, DropdownPosition } from './dropdownConstants';
+export { DropdownDirection, DropdownPosition, DropdownContext } from './dropdownConstants';
 export { default as DropdownGroup } from './DropdownGroup';
 export { default as DropdownItem } from './Item';
 export { default as DropdownSeparator } from './Separator';