+++ packages/patternfly-4/react-integration/demo-app-ts/src/components/demos/BackdropDemo/BackdropDemo.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Backdrop } from '@patternfly/react-core';
+
+export class BackdropDemo extends React.Component {
+  render() {
+    return (
+      <Backdrop className="backdrop" />
+    );
+  }
+}
+
+