+++ packages/patternfly-4/react-table/src/components/Table/index.js
@@ -1,3 +1,4 @@
+export { default as ActionsColumn } from './ActionsColumn';
 export { default as Table, TableContext, TableGridBreakpoint, TableVariant } from './Table';
 export { default as TableHeader } from './Header';
 export { default as TableBody } from './Body';