+++ packages/patternfly-4/react-integration/demo-app-ts/CHANGELOG.md
@@ -3,6 +3,138 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.22.1](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.22.0...demo-app-ts@1.22.1) (2019-06-14)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+# [1.22.0](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.21.5...demo-app-ts@1.22.0) (2019-06-14)
+
+
+### Features
+
+* **Accordion:** Accordion Component TypeScript Conversion ([#2121](https://github.com/patternfly/patternfly-react/issues/2121)) ([e2e981d](https://github.com/patternfly/patternfly-react/commit/e2e981d))
+
+
+
+
+
+## [1.21.5](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.21.4...demo-app-ts@1.21.5) (2019-06-14)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [1.21.4](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.21.3...demo-app-ts@1.21.4) (2019-06-14)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [1.21.3](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.21.2...demo-app-ts@1.21.3) (2019-06-13)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [1.21.2](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.21.1...demo-app-ts@1.21.2) (2019-06-13)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [1.21.1](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.21.0...demo-app-ts@1.21.1) (2019-06-12)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+# [1.21.0](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.20.0...demo-app-ts@1.21.0) (2019-06-12)
+
+
+### Features
+
+* **react-core:** allow dynamic rendering of anchor components ([#2117](https://github.com/patternfly/patternfly-react/issues/2117)) ([c1c3be9](https://github.com/patternfly/patternfly-react/commit/c1c3be9))
+
+
+
+
+
+# [1.20.0](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.19.2...demo-app-ts@1.20.0) (2019-06-11)
+
+
+### Features
+
+* **ApplicationLauncher:** Convert application launcher to typescript ([#2166](https://github.com/patternfly/patternfly-react/issues/2166)) ([9b90f94](https://github.com/patternfly/patternfly-react/commit/9b90f94))
+
+
+
+
+
+## [1.19.2](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.19.1...demo-app-ts@1.19.2) (2019-06-11)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [1.19.1](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.19.0...demo-app-ts@1.19.1) (2019-06-11)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+# [1.19.0](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.18.3...demo-app-ts@1.19.0) (2019-06-10)
+
+
+### Features
+
+* **backdrop:** Convert backdrop to typescript ([#2105](https://github.com/patternfly/patternfly-react/issues/2105)) ([fd3d771](https://github.com/patternfly/patternfly-react/commit/fd3d771))
+
+
+
+
+
+## [1.18.3](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.18.2...demo-app-ts@1.18.3) (2019-06-10)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [1.18.2](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.18.1...demo-app-ts@1.18.2) (2019-06-07)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
+## [1.18.1](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.18.0...demo-app-ts@1.18.1) (2019-06-06)
+
+**Note:** Version bump only for package demo-app-ts
+
+
+
+
+
 # [1.18.0](https://github.com/patternfly/patternfly-react/compare/demo-app-ts@1.17.0...demo-app-ts@1.18.0) (2019-06-05)